backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
djangorestframework==3.16.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
